Safari-609.1.11.1/LayoutTests/fast/canvas/pattern-too-large-to-create-expected.txt

CONSOLE MESSAGE: line 18: Canvas area exceeds the maximum limit (width * height > 268435456).
PASS: Saw exception.
